Refactored according to PR feedback
